# design/api/openapi.yaml
openapi: 3.1.0
info:
  title: Fiscal Note Workbench API
  version: 1.0.0
  description: |
    REST API for the DOR fiscal note proof of concept. Paths are grouped by module:
    identity, bills, search, notes, templates, workflow, notifications, reference, admin.
    Detailed request and response shapes are specified in design/research/*.md; this file
    is the route catalog and the shared schema skeleton the implementation must generate from
    or conform to. All paths are served under /api/v1.
servers:
  - url: /api/v1
security:
  - session: []
  - bearer: []
...
paths:
...
  /bills/{biennium}/{id}/notes:
    get:
      tags: [notes]
      summary: Note revisions on this bill visible to the caller, grouped by version
      parameters:
        - { $ref: "#/components/parameters/biennium" }
        - { $ref: "#/components/parameters/billId" }
      responses:
        "200": { description: List, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/NoteRevisionSummary" } } } } }

  # ---------------- search ----------------
...
  /notes:
    post:
      tags: [notes]
      summary: Create a note and its first revision for a bill version (or amendment)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [billKey, versionCode]
              properties:
                billKey: { type: string, example: "WA:2025-26:HB2402" }
                versionCode: { type: string, example: S }
                amendmentId: { type: string }
                kind: { type: string, enum: [note, estimate], default: note }
                templateId: { type: string }
                cloneFromRevisionId: { type: string, description: Copy the document from a prior revision (B.COM.28-29) }
                request:
                  type: object
                  properties:
                    requestId: { type: string, example: "2402-1-1" }
                    requestedAt: { type: string, format: date-time }
                    requestedBy: { type: string }
                    legContact: { type: object, properties: { name: { type: string }, phone: { type: string } } }
                    tenYearRequested: { type: boolean }
                confidential: { type: boolean, default: false }
                drafterId: { type: string }
                priority: { type: string, enum: [low, normal, high, urgent] }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/NoteRevisionSummary" } } } }
    get:
      tags: [notes]
      summary: List note revisions visible to the caller
      parameters:
        - { name: billKey, in: query, schema: { type: string } }
        - { name: state, in: query, schema: { type: string } }
        - { name: assignee, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer } }
        - { name: size, in: query, schema: { type: integer } }
      responses:
        "200": { description: List, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/NoteRevisionSummary" } } } } }
  /notes/{id}:
    get:
      tags: [notes]
      summary: Note revision summary (bill, version, request, workflow state, deadlines, assignees)
      parameters: [{ $ref: "#/components/parameters/noteRevisionId" }]
      responses:
        "200": { description: Summary, content: { application/json: { schema: { $ref: "#/components/schemas/NoteRevisionSummary" } } } }
    patch:
      tags: [notes]
      summary: Update metadata (confidential flag, priority, request fields, identifier override per B.RFA.03)
      parameters: [{ $ref: "#/components/parameters/noteRevisionId" }]
      requestBody: { content: { application/json: { schema: { type: object } } } }
      responses: { "200": { description: Updated } }
  /notes/{id}/revisions:
    post:
      tags: [notes]
      summary: Create the next revision for a new bill version or amendment, cloning the document
      parameters: [{ $ref: "#/components/parameters/noteRevisionId" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [versionCode]
              properties:
                versionCode: { type: string }
                amendmentId: { type: string }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/NoteRevisionSummary" } } } }
  /notes/{id}/document:
    get:
      tags: [notes]
      summary: Head document
      parameters: [{ $ref: "#/components/parameters/noteRevisionId" }]
      responses:
        "200":
          description: NoteDocument
          headers: { ETag: { schema: { type: string } } }
          content: { application/json: { schema: { $ref: "#/components/schemas/NoteDocument" } } }
    put:
      tags: [notes]
      summary: Save head document (autosave). Requires If-Match with the current version.
      parameters:
        - { $ref: "#/components/parameters/noteRevisionId" }
        - { name: If-Match, in: header, required: true, schema: { type: string } }
        - { name: force, in: query, schema: { type: boolean } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [doc, mode]
              properties:
                doc: { type: object, description: ProseMirror JSON }
                mode: { type: string, enum: [limited, full] }
                clientId: { type: string }
      responses:
        "200": { description: Saved, content: { application/json: { schema: { type: object, properties: { version: { type: integer }, savedAt: { type: string, format: date-time }, estimateData: { $ref: "#/components/schemas/EstimateData" }, validation: { $ref: "#/components/schemas/ValidationResult" } } } } } }
        "403": { description: Not editable by caller in the current workflow state }
        "412": { description: Version conflict; body carries the current head }
  /notes/{id}/versions:
    get:
      tags: [notes]
      summary: Document versions (autosave heads and named snapshots)
      parameters: [{ $ref: "#/components/parameters/noteRevisionId" }]
      responses:
        "200": { description: List, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/DocumentVersionSummary" } } } } }
    post:
      tags: [notes]
      summary: Name a snapshot of the head
      parameters: [{ $ref: "#/components/parameters/noteRevisionId" }]
      requestBody: { content: { application/json: { schema: { type: object, properties: { label: { type: string } } } } } }
      responses: { "201": { description: Created } }
  /notes/{id}/versions/{v}:
    get:
      tags: [notes]
      summary: A stored document version
      parameters:
        - { $ref: "#/components/parameters/noteRevisionId" }
        - { name: v, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: NoteDocument, content: { application/json: { schema: { $ref: "#/components/schemas/NoteDocument" } } } }
  /notes/{id}/versions/{v}/restore:
    post:
      tags: [notes]
      summary: Restore a version as the new head
      parameters:
        - { $ref: "#/components/parameters/noteRevisionId" }
        - { name: v, in: path, required: true, schema: { type: integer } }
      responses: { "201": { description: New head version } }
  /notes/{id}/diff:
    get:
      tags: [notes]
      summary: Redline between two document versions plus a table-cell diff
      parameters:
        - { $ref: "#/components/parameters/noteRevisionId" }
        - { name: from, in: query, required: true, schema: { type: integer } }
        - { name: to, in: query, required: true, schema: { type: integer } }
      responses:
        "200": { description: Diff, content: { application/json: { schema: { type: object, properties: { html: { type: string }, tables: { type: array, items: { type: object } } } } } } }
  /notes/{id}/lock:
    post:
      tags: [notes]
      summary: Acquire a soft edit lock
      parameters: [{ $ref: "#/components/parameters/noteRevisionId" }]
      responses:
        "200": { description: Held }
        "409": { description: Held by another user }
    delete:
      tags: [notes]
      summary: Release the lock
      parameters: [{ $ref: "#/components/parameters/noteRevisionId" }]
      responses: { "204": { description: Released } }
  /notes/{id}/comments:
    get:
      tags: [notes]
      summary: Comment threads
      parameters: [{ $ref: "#/components/parameters/noteRevisionId" }]
      responses:
        "200": { description: Threads, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/CommentThread" } } } } }
    post:
      tags: [notes]
      summary: Open a thread anchored to a range
      parameters: [{ $ref: "#/components/parameters/noteRevisionId" }]
      requestBody: { content: { application/json: { schema: { type: object, required: [anchorText, body], properties: { anchorText: { type: string }, body: { type: string } } } } } }
      responses: { "201": { description: Created } }
  /notes/{id}/comments/{cid}:
    patch:
      tags: [notes]
      summary: Resolve or reopen a thread
      parameters:
        - { $ref: "#/components/parameters/noteRevisionId" }
        - { name: cid, in: path, required: true, schema: { type: string } }
      requestBody: { content: { application/json: { schema: { type: object, properties: { status: { type: string, enum: [open, resolved] } } } } } }
      responses: { "200": { description: Updated } }
    delete:
      tags: [notes]
      summary: Delete a thread (author or reviewer)
      parameters:
        - { $ref: "#/components/parameters/noteRevisionId" }
        - { name: cid, in: path, required: true, schema: { type: string } }
      responses: { "204": { description: Deleted } }
  /notes/{id}/comments/{cid}/messages:
    post:
      tags: [notes]
      summary: Reply in a thread
      parameters:
        - { $ref: "#/components/parameters/noteRevisionId" }
        - { name: cid, in: path, required: true, schema: { type: string } }
      requestBody: { content: { application/json: { schema: { type: object, required: [body], properties: { body: { type: string } } } } } }
      responses: { "201": { description: Created } }
  /notes/{id}/export:
    post:
      tags: [notes]
      summary: Export a document version as docx, pdf, xml (FNS placeholder), or html
      parameters:
        - { $ref: "#/components/parameters/noteRevisionId" }
        - { name: format, in: query, required: true, schema: { type: string, enum: [docx, pdf, xml, html] } }
        - { name: version, in: query, schema: { type: integer, description: "Defaults to the approved snapshot for viewers, the head for editors" } }
        - { name: comments, in: query, schema: { type: boolean, default: false } }
      responses:
        "200": { description: File }
        "202": { description: Job accepted, content: { application/json: { schema: { $ref: "#/components/schemas/JobAccepted" } } } }
        "422": { description: Validation failed (totals do not reconcile, required slots empty) }
  /notes/{id}/validate:
    get:
      tags: [notes]
      summary: Validate the head document (required slots, table reconciliation)
      parameters: [{ $ref: "#/components/parameters/noteRevisionId" }]
      responses:
        "200": { description: ValidationResult, content: { application/json: { schema: { $ref: "#/components/schemas/ValidationResult" } } } }
  /export-jobs/{jobId}:
    get:
      tags: [notes]
      summary: Export job status
      parameters: [{ name: jobId, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Status, content: { application/json: { schema: { type: object, properties: { status: { type: string }, url: { type: string } } } } } }

  # ---------------- templates ----------------
...
components:
  securitySchemes:
    session: { type: apiKey, in: cookie, name: session }
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    biennium: { name: biennium, in: path, required: true, schema: { type: string, pattern: "^\\d{4}-\\d{2}$" }, example: "2025-26" }
    billId: { name: id, in: path, required: true, schema: { type: string, pattern: "^[HS](B|JR|JM|CR|R|I)\\d{1,5}$" }, example: HB2402 }
...
    noteRevisionId: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
  schemas:
...
    NoteRevisionSummary:
      type: object
      properties:
        noteRevisionId: { type: string }
        noteId: { type: string }
        billKey: { type: string }
        versionCode: { type: string }
        versionLabel: { type: string }
        amendmentId: { type: string }
        kind: { type: string, enum: [note, estimate] }
        requestId: { type: string }
        confidential: { type: boolean }
        priority: { type: string }
        state: { type: string }
        drafterStatus: { type: string }
        reviewerStatus: { type: string }
        drafter: { type: object }
        reviewer: { type: object }
        deadlines: { type: array, items: { type: object, properties: { kind: { type: string }, dueAt: { type: string, format: date-time } } } }
        effectiveDueAt: { type: string, format: date-time }
        headVersion: { type: integer }
        approvedVersion: { type: integer }
        previousRevisionId: { type: string }
        supersededBy: { type: string }
        updatedAt: { type: string, format: date-time }
    NoteDocument:
      type: object
      properties:
        noteId: { type: string }
        version: { type: integer }
        mode: { type: string, enum: [limited, full] }
        doc: { type: object, description: ProseMirror JSON }
        templateId: { type: string }
        templateVersion: { type: integer }
        updatedAt: { type: string, format: date-time }
        updatedBy: { type: string }
    DocumentVersionSummary:
      type: object
      properties:
        version: { type: integer }
        label: { type: string }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        summary: { type: string }
    EstimateData:
      type: object
      description: Extracted from estimateTable nodes; six fiscal-year values per row, biennia derived
      properties:
        revenue: { type: array, items: { $ref: "#/components/schemas/EstimateRow" } }
        expenditure: { type: array, items: { $ref: "#/components/schemas/EstimateRow" } }
        fte: { type: array, items: { $ref: "#/components/schemas/EstimateRow" } }
        local: { type: array, items: { $ref: "#/components/schemas/EstimateRow" } }
        tenYear: { type: array, items: { type: object } }
        flags: { type: object, additionalProperties: { type: boolean } }
    EstimateRow:
      type: object
      properties:
        key: { type: string }
        fund: { type: string }
        fundName: { type: string }
        source: { type: string }
        fy: { type: object, additionalProperties: { type: number } }
        biennia: { type: object, additionalProperties: { type: number } }
    ValidationResult:
      type: object
      properties:
        ok: { type: boolean }
        errors: { type: array, items: { type: object, properties: { code: { type: string }, path: { type: string }, message: { type: string } } } }
        warnings: { type: array, items: { type: object } }
        unfilledSlots: { type: array, items: { type: string } }
    CommentThread:
      type: object
      properties:
        id: { type: string }
        status: { type: string, enum: [open, resolved] }
        anchorText: { type: string }
        detached: { type: boolean }
        position: { type: integer }
        messages: { type: array, items: { type: object, properties: { id: { type: string }, authorId: { type: string }, authorName: { type: string }, body: { type: string }, createdAt: { type: string, format: date-time } } } }
...
    JobAccepted:
      type: object
      properties:
        job_id: { type: string }
        status_url: { type: string }